achieve 100% test coverage
This seeks to establish 100% test coverage by homing `run`'s testable
logic in `run.js` and making `index.js` a logic-less invocation of
`run`.

```
-----------|---------|----------|---------|---------|-------------------
File       | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
-----------|---------|----------|---------|---------|-------------------
All files  |     100 |      100 |     100 |     100 |
 github.js |     100 |      100 |     100 |     100 |
 run.js    |     100 |      100 |     100 |     100 |
 utils.js  |     100 |      100 |     100 |     100 |
-----------|---------|----------|---------|---------|-------------------

Test Suites: 3 passed, 3 total
Tests:       21 passed, 21 total
Snapshots:   0 total
Time:        0.855 s, estimated 1 s
Ran all test suites.
✨  Done in 1.71s.
```
